mcp-implementation/pyproject.toml
[project]
name = "mcp-implementation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-mcp-adapters>=0.1.11",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "python-dotenv>=1.1.1",
]
